requirements.txt

Flask==2.3.2
Flask-Cors==3.0.10
PyMySQL==1.0.2
python-dotenv==1.0.0


